+++ Flashcards/Views/DeckView.swift
@@ -11,14 +11,16 @@ import Combine
 class DeckViewViewModel: ObservableObject {
     @Published var progress: Double = 0
     @Published var flashcardModel: FlashcardModel?
+    @Published var didFinish = false
     
-//    private var cancellable: AnyCancellable?
     private let flashcardModels: [FlashcardModel]
     private var index: Int = 0 {
         didSet {
             progress = Double(index) / Double(flashcardModels.count) * 100
         }
     }
+    
+    //    var onFinish: (() -> Void)?
     
     init(flashcardModels: [FlashcardModel]) {
         self.flashcardModels = flashcardModels
@@ -33,27 +35,37 @@ class DeckViewViewModel: ObservableObject {
         }
         else {
             print("finished")
-//            present well done view
+            //            onFinish?()
+            didFinish = true
         }
     }
 }
 
 struct DeckView: View {
     @ObservedObject var viewModel: DeckViewViewModel
+    @Environment(\.presentationMode) var mode: Binding<PresentationMode>
     
     var body: some View {
-        return VStack {
-            if let flashcardModel = viewModel.flashcardModel {
-                FlashcardView(model: flashcardModel) {
-                    viewModel.cardViewed($0, model: $1)
-                }.id(flashcardModel.id)
-                
-                ProgressBarView(progress: viewModel.progress)
-                    .frame(width: 400, height: 30)
-            }
-            else {
-                // later well done view
-                EmptyView()
+        ZStack {
+            if viewModel.didFinish {
+                Button("Well Done") {
+                    self.mode.wrappedValue.dismiss()
+                }
+            } else {
+                VStack {
+                    if let flashcardModel = viewModel.flashcardModel {
+                        FlashcardView(model: flashcardModel) {
+                            viewModel.cardViewed($0, model: $1)
+                        }.id(flashcardModel.id)
+                        
+                        ProgressBarView(progress: viewModel.progress)
+                            .frame(width: 400, height: 30)
+                    }
+                    else {
+                        // later well done view
+                        EmptyView()
+                    }
+                }
             }
         }
     }
